review

+++ src/backend/PeopleHub.Application/Models/FriendInfo.cs
@@ -2,5 +2,5 @@ using PeopleHub.Domain.Enums;
 
 namespace PeopleHub.Application.Models;
 
-public sealed record FriendInfo(int Id, string Name, string Surname, int Age, string City, Gender Gender, string Bio,
+public sealed record FriendInfo(long Id, string Name, string Surname, int Age, string City, Gender Gender, string Bio,
     FriendRequestStatus Status);
