+++ src/backend/PeopleHub.Application/Models/SearchedUser.cs
@@ -1,3 +1,3 @@
 namespace PeopleHub.Application.Models;
 
-public sealed record SearchedUser(int Id, string FirstName, string SecondName, string City, string Biography);
+public sealed record SearchedUser(long Id, string FirstName, string SecondName, string City, string Biography);
