+++ src/backend/PeopleHub.Domain/Entities/Account.cs
@@ -2,13 +2,13 @@ using PeopleHub.Domain.ValueObjects;
 
 namespace PeopleHub.Domain.Entities;
 
-public sealed class Account(int id, Email email, PasswordHash password, int userId)
+public sealed class Account(long id, Email email, PasswordHash password, long userId)
 {
-    public int Id { get; } = id;
+    public long Id { get; } = id;
     public Email Email { get; } = email;
     public PasswordHash Password { get; } = password;
-    public int UserId { get; } = userId;
+    public long UserId { get; } = userId;
 
-    public static Account Create(Email email, PasswordHash password, int userId) =>
+    public static Account Create(Email email, PasswordHash password, long userId) =>
         new(0, email, password, userId);
 }
