+++ src/backend/PeopleHub.Domain/Repositories/IUserRepository.cs
@@ -6,7 +6,7 @@ public interface IUserRepository
 {
     Task<User?> GetAsync(long id, CancellationToken cancellationToken = default);
 
-    Task<int?> CreateAsync(User user, CancellationToken cancellationToken = default);
+    Task<long?> CreateAsync(User user, CancellationToken cancellationToken = default);
 
     Task UpdateAsync(User user, CancellationToken cancellationToken = default);
 }
