+++ src/backend/PeopleHub.Infrastructure/Repositories/AccountRepository.cs
@@ -8,7 +8,7 @@ namespace PeopleHub.Infrastructure.Repositories;
 
 internal class AccountRepository(DbClient dbClient) : IAccountRepository
 {
-    public async Task<int> CreateAsync(Account account, CancellationToken cancellationToken = default)
+    public async Task<long> CreateAsync(Account account, CancellationToken cancellationToken = default)
     {
         var scalar = await dbClient.ExecuteScalarAsync(
             $"INSERT INTO {DbClient.AccountsTable} (email, password, user_id) " +
@@ -19,7 +19,7 @@ internal class AccountRepository(DbClient dbClient) : IAccountRepository
                 ("userId", account.UserId)
             ]);
 
-        return Convert.ToInt32(scalar);
+        return Convert.ToInt64(scalar);
     }
 
     public async Task<bool> ExistsAsync(Email email, CancellationToken cancellationToken = default)
@@ -59,10 +59,10 @@ internal class AccountRepository(DbClient dbClient) : IAccountRepository
 
         var dataRow = dataTable.Rows[0];
         return new Account(
-            Convert.ToInt32(dataRow["id"]),
+            Convert.ToInt64(dataRow["id"]),
             Email.Create(dataRow["email"].ToString()),
             new PasswordHash(dataRow["password"].ToString()),
-            Convert.ToInt32(dataRow["user_id"])
+            Convert.ToInt64(dataRow["user_id"])
         );
     }
 }
